requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
configparser==4.0.2
python-telegram-bot==12.4.2
requests==2.18.4
requests-unixsocket==0.1.5
virtualenv==20.0.2
web.py==0.40
